FGIsolationTool/requirements.txt
tkinter
Pillow
rembg
customtkinter